user withdraw completion
